Adjust component name, location and adjust for new data structure
Now it is possible to view the person page with the new data
structure.
