--- src/components/organisms/profiles/index.js
@@ -1,16 +1,0 @@
-//#region > Imports
-//> Components
-// Import all components to export them for easy access from parent components
-import SoftwareEngineer from "./SoftwareEngineer";
-//#endregion
-
-//#region > Exports
-//> Components
-// Export the components for easy access from parent components
-export { SoftwareEngineer };
-//#endregion
-
-/**
- * SPDX-License-Identifier: (EUPL-1.2)
- * Copyright © 2019-2020 Simon Prast
- */
