+++ src/components/organisms/sections/person/InfoCard/infocard.scss
