+++ src/components/organisms/sections/person/Tabs/index.jsx
@@ -11,9 +11,9 @@ import { MDBBadge } from "mdbreact";
 import { connect } from "react-redux";
 
 //> Components
-import { ProjectTab, OverviewTab, TalksTab } from "../tabs";
+import { ProjectTab, OverviewTab, TalksTab } from "../../../tabs";
 //> Style sheet
-import "./softwaretabs.scss";
+import "./tabs.scss";
 //#endregion
 
 //#region > Components
@@ -30,13 +30,13 @@ class SoftwareTabs extends React.Component {
   };
 
   isSameOrigin = () => {
-    const { fetchedUser, loggedUser } = this.props;
+    const { fetchedPerson, loggedUser } = this.props;
 
-    return fetchedUser.username === loggedUser.username;
+    return fetchedPerson.username === loggedUser.username;
   };
 
   render() {
-    const { fetchedUser } = this.props;
+    const { fetchedPerson } = this.props;
     const { activeTab } = this.state;
     const tabItems = [
       {
@@ -48,8 +48,8 @@ class SoftwareTabs extends React.Component {
       {
         title: "Projects",
         visible: true,
-        pill: this.props.fetchedUser?.platformData?.profile?.repositories
-          ? this.props.fetchedUser?.platformData?.profile?.repositories?.length
+        pill: fetchedPerson?.person?.repositories
+          ? fetchedPerson?.person?.repositories.length
           : "0",
         notification: false,
       },
@@ -74,9 +74,7 @@ class SoftwareTabs extends React.Component {
         title: "Talks",
         visible: true,
         notification: false,
-        pill: this.props.fetchedUser?.platformData.talks
-          ? this.props.fetchedUser?.platformData.talks.length
-          : "0",
+        pill: fetchedPerson?.talks ? fetchedPerson?.talks.length : "0",
         notification: false,
       },
     ];
@@ -100,15 +98,18 @@ class SoftwareTabs extends React.Component {
         </ul>
         <div className="p-3 content">
           {activeTab === 0 && (
-            <OverviewTab
-              platformData={fetchedUser && fetchedUser.platformData}
-              sameOrigin={this.isSameOrigin()}
-            />
+            // <OverviewTab
+            //   platformData={fetchedUser && fetchedUser.platformData}
+            //   sameOrigin={this.isSameOrigin()}
+            // />
+            <h1></h1>
           )}
           {activeTab === 1 && (
             <ProjectTab
               repoList={
-                fetchedUser && fetchedUser.platformData.profile.repositories
+                fetchedPerson?.person.repositories
+                  ? fetchedPerson.person.repositories
+                  : []
               }
             />
           )}
@@ -117,7 +118,11 @@ class SoftwareTabs extends React.Component {
               This feature is not available just yet.
             </p>
           )}
-          {activeTab === 5 && <TalksTab {...this.props} />}
+          {activeTab === 5 && (
+            <TalksTab
+              talkList={fetchedPerson?.talks ? fetchedPerson.talks : []}
+            />
+          )}
         </div>
       </div>
     );
@@ -127,8 +132,8 @@ class SoftwareTabs extends React.Component {
 
 //#region > Redux Mapping
 const mapStateToProps = (state) => ({
-  loggedUser: state.auth.loggedUser,
-  fetchedUser: state.user.fetchedUser,
+  loggedUser: state.user.user,
+  fetchedPerson: state.general.fetchedPerson,
 });
 
 const mapDispatchToProps = (dispatch) => {
