+++ src/components/organisms/sections/person/Tabs/tabs.scss
