+++ src/components/pages/PersonPage/person.scss
