+++ src/components/pages/index.js
@@ -3,7 +3,7 @@
 // Import all components to export them for easy access from parent components
 import CompanyPage from "./CompanyPage";
 import HomePage from "./HomePage";
-import ProfilePage from "./ProfilePage";
+import PersonPage from "./PersonPage";
 import TalkPage from "./TalkPage";
 import SettingsPage from "./SettingsPage";
 // The temp page displays sections that can not be displayed in profile
@@ -12,7 +12,7 @@ import TempPage from "./TempPage";
 
 //#region > Exports
 //> Pages
-export { CompanyPage, HomePage, ProfilePage, TalkPage, SettingsPage, TempPage };
+export { CompanyPage, HomePage, PersonPage, TalkPage, SettingsPage, TempPage };
 //#endregion
 
 /**
